feat: [NUTR-104] - fixes test
